Projects/hello-world/requirements.txt
blessed==1.20.0
botocore==1.36.4
certifi==2024.8.30
chalice==1.31.3
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
editor==1.6.6
git-filter-repo==2.38.0
httpie==3.2.4
idna==3.10
inquirer==3.4.0
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
Pillow==10.1.0
pinterest-api-sdk==0.2.1
Pinterest-Generated-Client==0.1.7
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
runs==1.2.2
setuptools==75.8.0
six==1.16.0
tqdm==4.66.1
typing==3.7.4.3
urllib3==1.26.12
wcwidth==0.2.13
wheel==0.45.1
xmod==1.8.1
